Tiny refactor of existing Mastodon instance tests
